flavours: Remove infra and compute nodes
Since infra and compute nodes are now calculated based on a combination
of CCS and MultiAZ flags, they are no longer considered part of
a flavour, since they can be different defaults in each combination of
the flags. We now calculate the required default nodes as part of the
cluster creation process.