sts: Allow adding/removing operator roles
This change allows users to fetch, add, remove operator roles. Operator
roles contain the name and namespace where the manifest configmap will
set it, along with a service account and associated role ARN for
verifying authentication.

This allows for day-2 operators and addons that rely on AssumeRole calls
for accessing cloud resources.

https://issues.redhat.com/browse/SDA-5366