OCM-209 | feat: Add cluster autoscaler structs
This adds a reduced part of cluster autoscaler parameters (the ones
prioritized by PMs). We're intending in adding the rest of the
parameters after we see how it goes (in addition to managing it in small
chunks of code, as the CA object has lots of options).

You can see the intended logic in clusters-service PR #5798. The
ClusterAutoscaler is part of cluster endpoints' payload and doesn't have
dedicated endpoints. For example, you can do:
```
$ ocm patch /api/clusters_mgmt/v1/clusters/{cluster_id} << EOF
{
	"autoscaler": {
		"scale_down": {
			"delay_after_add": "3h"
		}
	}
}
EOF
```

but you *cannot* do:
```
$ ocm get /api/clusters_mgmt/v1/clusters/{cluster_id}/autoscalers
```

or anything alike.