remove referencing in models/aro_hcp
We achieve that by removing any referencing via the
ref statement in the models/aro_hcp definitions. We then
copy all the attributes for the resources we have defined
in the existing aro-hcp resources from model/clusters_mgmt,
even if not currently used by ARO-HCP.

However, we remove any locator statements to avoid providing
additional endpoints under the aro_hcp/v1alpha1 API URL for
the existing resources defined in aro-hcp.

The copying of all elements is needed to resolve all the attribute
references now that we removed the referencing mechanism. It also
ensures that there are no dependencies with ocm commercial from within
the aro-hcp models.

This commit does not copy resources that might be referenced by
the copied attributes. This will be done in a follow-up commit.

This will end up adding many data types in the generated Go
models that won't be used and they will be available as data types.
This will also modify existing methods of the generated Go types
to use the duplicated data types. This can have an impact
on the consumers of the models but for aro-hcp case we expect
if they need to change it should not be a big amount of work.o