remove resource types for aro-hcp models we do not want to generate clients
Some of the models we duplicated from ocm commercial into
aro-hcp are not currently used in the aro-hcp offering. We
remove the `resource` definition for them. This ensures
that the OCM SDK Go will not generate Go code to interact
with the endpoints associated to these resources.
The class and struct associated models will be kept so
the Go data types themselves will still be generated. We do
the latter because aro-hcp models that are currently in use
might reference those.

We performed this by copying the list of files containing
_resource in their name from the ocm commercial models
into aro-hcp models, excluding the resources files that
already existed in aro-hcp models before the split/duplication
occurred.