aro-hcp: update some code comments mentioning clusters_mgmt to aro_hcp
To be accurate with what's implemented.

In model/aro_hcp there are still some references to clusters_mgmt in
some example URLs but that is intentional because the endpoints
themselves in CS are only offered under clusters_mgmt for those.

Further cleanup and modifications to make it accurate can be done in
the future if further changes are performed there.